requirements.txt
Django==4.2
django-allauth==0.58.2
supabase==1.0.3
python-dotenv==1.0.0
django-ratelimit==3.0.1
django-reversion==5.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
whitenoise==6.6.0

# Chatbot integration requirements
llama-index==0.9.2
openai==1.3.0
nltk==3.8.1
python-dateutil==2.8.2
selenium==4.14.0
dnspython==2.4.2
